requirements.txt
absl-py                      2.1.0
altair                       5.5.0
anyio                        4.6.2
argon2-cffi                  21.3.0
argon2-cffi-bindings         21.2.0
asttokens                    2.0.5
astunparse                   1.6.3
async-lru                    2.0.4
attrs                        24.3.0
babel                        2.16.0
beautifulsoup4               4.12.3
bleach                       6.2.0
blinker                      1.9.0
Brotli                       1.0.9
cachetools                   5.5.2
certifi                      2025.1.31
cffi                         1.17.1
charset-normalizer           3.3.2
click                        8.1.8
colorama                     0.4.6
comm                         0.2.1
debugpy                      1.8.11
decorator                    5.1.1
defusedxml                   0.7.1
executing                    0.8.3
fastjsonschema               2.20.0
flatbuffers                  25.2.10
gast                         0.6.0
gitdb                        4.0.12
GitPython                    3.1.44
google-pasta                 0.2.0
grpcio                       1.71.0
h11                          0.14.0
h5py                         3.13.0
httpcore                     1.0.2
httpx                        0.27.0
idna                         3.10
ipykernel                    6.29.5
ipython                      8.30.0
jedi                         0.19.2
Jinja2                       3.1.5
joblib                       1.4.2
json5                        0.9.25
jsonschema                   4.23.0
jsonschema-specifications    2023.7.1
jupyter_client               8.6.3
jupyter_core                 5.7.2
jupyter-events               0.12.0
jupyter-lsp                  2.2.0
jupyter_server               2.15.0
jupyter_server_terminals     0.4.4
jupyterlab                   4.3.4
jupyterlab_pygments          0.3.0
jupyterlab_server            2.27.3
keras                        3.9.0
libclang                     18.1.1
Markdown                     3.7
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
matplotlib-inline            0.1.6
mdurl                        0.1.2
mistune                      3.1.2
ml-dtypes                    0.4.1
namex                        0.0.8
narwhals                     1.30.0
nbclient                     0.10.2
nbconvert                    7.16.6
nbformat                     5.10.4
nest-asyncio                 1.6.0
notebook                     7.3.2
notebook_shim                0.2.3
numpy                        2.0.2
opt_einsum                   3.4.0
optree                       0.14.1
overrides                    7.4.0
packaging                    24.2
pandas                       2.2.3
pandocfilters                1.5.0
parso                        0.8.4
pickle-mixin                 1.0.2
pillow                       11.1.0
pip                          25.0
platformdirs                 3.10.0
prometheus_client            0.21.0
prompt-toolkit               3.0.43
protobuf                     5.29.3
psutil                       5.9.0
pure-eval                    0.2.2
pyarrow                      19.0.1
pycparser                    2.21
pydeck                       0.9.1
Pygments                     2.15.1
PySocks                      1.7.1
python-dateutil              2.9.0.post0
python-json-logger           3.2.1
pytz                         2025.1
pywin32                      308
pywinpty                     2.0.14
PyYAML                       6.0.2
pyzmq                        26.2.0
referencing                  0.30.2
requests                     2.32.3
rfc3339-validator            0.1.4
rfc3986-validator            0.1.1
rich                         13.9.4
rpds-py                      0.22.3
scikit-learn                 1.6.1
scipy                        1.15.2
Send2Trash                   1.8.2
setuptools                   75.8.0
six                          1.16.0
smmap                        5.0.2
sniffio                      1.3.0
soupsieve                    2.5
stack-data                   0.2.0
streamlit                    1.43.1
tenacity                     9.0.0
tensorboard                  2.18.0
tensorboard-data-server      0.7.2
tensorflow                   2.18.0
tensorflow_intel             2.18.0
tensorflow-io-gcs-filesystem 0.31.0
termcolor                    2.5.0
terminado                    0.17.1
threadpoolctl                3.5.0
tinycss2                     1.4.0
toml                         0.10.2
tornado                      6.4.2
traitlets                    5.14.3
typing_extensions            4.12.2
tzdata                       2025.1
urllib3                      2.3.0
watchdog                     6.0.0
wcwidth                      0.2.5
webencodings                 0.5.1
websocket-client             1.8.0
Werkzeug                     3.1.3
wheel                        0.45.1
win-inet-pton                1.1.0
wrapt                        1.17.2
